requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0

# ML / Embedding
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
torch==2.2.2
transformers==4.40.0

# Dataset
datasets==2.19.0
huggingface-hub==0.23.0

# Utilities
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7

# Benchmarking / Load testing
locust==2.28.0
